requirements.txt
numpy
matplotlib
seaborn
scikit-learn
tensorflow-macos
tensorflow-metal
